backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.35
alembic==1.13.0
asyncpg==0.30.0
psycopg2-binary==2.9.9
python-docx==1.1.0
striprtf==0.0.26
PyPDF2==3.0.1
anthropic==0.39.0
pydantic==2.9.0
pydantic-settings==2.5.0
redis==5.1.0
celery==5.4.0
boto3==1.35.0
python-dotenv==1.0.0
httpx==0.27.0
aiofiles==24.1.0
stripe==7.0.0
